requirements.txt
Flask==1.00
tweepy==3.6.0
google-cloud-language==1.0.1
gunicorn==19.7.1
